+++ src/mesa/swrast_setup/ss_tritmp.h
@@ -37,7 +37,7 @@ static void TAG(triangle)(GLcontext *ctx, GLuint e0, GLuint e1, GLuint e2 )
    GLuint facing = 0;
    GLchan saved_color[3][4];
    GLchan saved_spec[3][4];
-   GLuint saved_index[3];
+   GLfloat saved_index[3];
 
    v[0] = &verts[e0];
    v[1] = &verts[e1];
